Grid Flow v2
Experimenting with ebiten for drawing, rather than printing to a
terminal.

Results thus far: this is waaaaaay faster than printing to my terminal.
Guess that's much slower than I thought, huh.

Obviously the code here is very rough round the edges, but as a proof of
concept, it's behaving roughly how I want.