Initial version of this which mostly works
Started out as a straight copy of github.com/sahilm/fuzzy/_example,
which I have modified. As such, a lot of the code makes no sense in this
context. It needs refactoring.

But as an initial proof of concept, it behaves correctly. Mostly.

The last feature that needs to be added for it to behave as expected is
to be able to select a number from the list.