Minor optimisations on boids
Firstly, make the sepration/alignment/cohesion multiplers config
variables. Not an optimisation as such, but a useful feature.

Add a worker pool to update boids concurrently. Useful at the moment,
with our expensive neighbourhood calculations.

Consider horizontal/vertical distance from a potential neighbour before
calculating the actual distance. The squareroot function used in
calculating geometric distance between two points is an expensive
operation. No need to actually run it if we can run a cheaper comparison
first.

This is still nowhere near as efficient as a QuadTree would be, as we
are still comparing ourselves to every other boid. But it's a little bit
better.