Ebiten Keyboard InputHack experiment, for mobile JS
Run natively, this just uses the keyboard for input.

Run in JS on desktop, you can just use keyboard normally too.

On mobile, you can't, so this includes two workarounds.

The first uses JavaScript prompt() to get text.
This works fine, but isn't particularly great.

The second adds an inputbox to the DOM, and focuses it.
This works fine in JS on Desktop, but not on iOS.
It may be possbile to tweak it to work on iOS, but I'm not sure.