Workaround for inputHack for mobile
To focus the inputHack text area on mobile, we need to do it as the
result of the user doing _something_

For this simple experiment, I'm just setting that to be a "tap anywhere
on the canvas". In a real game, this would be something like "tap
anywhere on the canvas, then only focus based on whether or not we
tapped within the bounds of an in-game text area.